requirements.txt
numpy==1.22.4
pandas==1.4.3
rdkit-pypi==2022.3.5
scikit-learn==1.1.1
tensorboard==2.9.1
torch==1.12.0
torch-cluster==1.6.0
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.14
torch-spline-conv==1.2.1



